get href & whole_url

+++ multi_down.sh
@@ -2,8 +2,15 @@
 
 link="${1:?link is required!}"
 
+echo $link
+if [ "$link" == "" ];
+then
+   exit 2 
+fi
+
 file="index.html"
 bak_file="bak_$file"
+i=0
 
 # rm all html
 rm -f *.html
@@ -11,8 +18,17 @@ rm -f *.html
 # get index.html
 wget $link -O $file
 
-cat $file | grep "<a" > $bak_file
+cat $file > $bak_file
 
 sed -i -e 's/</\n</g' $bak_file
 sed -i -e 's/>/\n>/g' $bak_file
 
+url_list=`cat $bak_file | grep "href*=*\"" | grep "<a" | cut -d "\"" -f 2 | grep "/$" | grep -v "\.\."`
+
+for url in $url_list
+do
+    whole_url_list[$i]="$link$url"
+    i=`expr $i+1`
+done
+
+echo ${whole_url_list[@]}
